Factor create button out of providers page
